requirement.txt
pandas
sklearn
matplotlib
numpy 
torch
transformer
tensorflow
keras
   